refactor: move all command files into src/commands/
Move fleet, wake, pulse, done, oracle, overview, fleet-init into
commands/ directory. Update all import paths. Matches existing
pattern (comm.ts, view.ts, completions.ts already in commands/).
